Classification/code/requirements.txt
monai 
itk
pillow
pandas
pydicom
imagecodecs
